# dist/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Caltex Music",
        "description": "",
        "contact": {
            "email": "tauovir.khan@appinventiv.com"
        },
        "version": "1.0"
    },
    "host": "caltex.appinventive.com",
    "basePath": "/api/v1",
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/country-list": {
            "get": {
                "tags": [
                    "Countries"
                ],
                "summary": "Get All active country",
                "description": "Get All active country",
                "operationId": "Country",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Basic Auth",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully executed"
                    },
                    "201": {
                        "description": "fail to execute"
                    },
                    "202": {
                        "description": "Require parameter missing or problem"
                    },
                    "203": {
                        "description": "Basic Authentication fail"
                    },
                    "500": {
                        "description": "Internal server Error "
                    }
                },
                "security": [
                    {
                        "basicAuth": [
                            "Username: ",
                            "Password: "
                        ]
                    }
                ]
            }
        },
        "/update-profile": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Profile Update",
                "description": "Profile update",
                "operationId": "Profile Update",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Basic Auth",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "accessToken",
                        "in": "header",
                        "description": "AccessToken",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "cust_dob",
                        "in": "formData",
                        "description": "Date of birth",
                        "type": "string"
                    },
                    {
                        "name": "cust_phone",
                        "in": "formData",
                        "description": "Phone",
                        "type": "string"
                    },
                    {
                        "name": "country_id",
                        "in": "formData",
                        "description": "Country Id",
                        "type": "number"
                    },
                    {
                        "name": "cust_state",
                        "in": "formData",
                        "description": "State",
                        "type": "string"
                    },
                    {
                        "name": "cust_city",
                        "in": "formData",
                        "description": "City ",
                        "type": "number"
                    },
                    {
                        "name": "cust_gender",
                        "in": "formData",
                        "description": "Gender",
                        "type": "number"
                    },
                    {
                        "name": "cust_zipcode",
                        "in": "formData",
                        "description": "Zipcode",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully executed"
                    },
                    "201": {
                        "description": "fail to execute"
                    },
                    "202": {
                        "description": "Require parameter missing or problem"
                    },
                    "204": {
                        "description": "Api token invalid"
                    },
                    "203": {
                        "description": "Basic Authentication fail"
                    },
                    "205": {
                        "description": "Your account is in active"
                    },
                    "206": {
                        "description": "Your account has been blocked"
                    },
                    "500": {
                        "description": "Internal server Error "
                    }
                },
                "security": [
                    {
                        "basicAuth": [
                            "Username: ",
                            "Password: "
                        ],
                        "Api token": [
                            "accessToken: ",
                            "***********: "
                        ]
                    }
                ]
            }
        },
        "/change-password": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Change passwrod",
                "description": "Change password",
                "operationId": "Change password",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Basic Auth",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "accessToken",
                        "in": "header",
                        "description": "AccessToken",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "old_password",
                        "in": "formData",
                        "description": "Country Id",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "new_password",
                        "in": "formData",
                        "description": "State",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "repaet_password",
                        "in": "formData",
                        "description": "Gender",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully executed"
                    },
                    "201": {
                        "description": "fail to execute"
                    },
                    "202": {
                        "description": "Require parameter missing or problem"
                    },
                    "204": {
                        "description": "Api token invalid"
                    },
                    "203": {
                        "description": "Basic Authentication fail"
                    },
                    "205": {
                        "description": "Your account is in active"
                    },
                    "206": {
                        "description": "Your account has been blocked"
                    },
                    "500": {
                        "description": "Internal server Error "
                    }
                },
                "security": [
                    {
                        "basicAuth": [
                            "Username: ",
                            "Password: "
                        ],
                        "Api token": [
                            "accessToken: ",
                            "***********: "
                        ]
                    }
                ]
            }
        },
        "/logout": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "logout",
                "description": "logout",
                "operationId": "logout",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Basic Auth",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "accessToken",
                        "in": "header",
                        "description": "User Access Token for accessing Api",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully executed"
                    },
                    "201": {
                        "description": "fail to execute"
                    },
                    "202": {
                        "description": "Require parameter missing or problem"
                    },
                    "204": {
                        "description": "Api token invalid"
                    },
                    "203": {
                        "description": "Basic Authentication fail"
                    },
                    "205": {
                        "description": "Your account is in active"
                    },
                    "206": {
                        "description": "Your account has been blocked"
                    },
                    "500": {
                        "description": "Internal server Error "
                    }
                },
                "security": [
                    {
                        "basicAuth": [
                            "Username: ",
                            "Password: "
                        ]
                    }
                ]
            }
        },
...
        "/language": {
            "get": {
                "tags": [
                    "Languages"
                ],
                "summary": "Get All active Languages",
                "description": "Get All active country",
                "operationId": "Country",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Basic Auth",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully executed"
                    },
                    "201": {
                        "description": "fail to execute"
                    },
                    "202": {
                        "description": "Require parameter missing or problem"
                    },
                    "203": {
                        "description": "Basic Authentication fail"
                    },
                    "500": {
                        "description": "Internal server Error "
                    }
                },
                "security": [
                    {
                        "basicAuth": [
                            "Username: ",
                            "Password: "
                        ]
                    }
                ]
            }
        }
    },
...
    "securityDefinitions": {
        "basicAuth": {
            "type": "basic"
        }
    }
}